test/requirements.txt
pytest
pytest-asyncio
pytest-aiohttp